feat: implement getting transactions by user

+++ src/transactions/entities/transaction.entity.ts
@@ -13,7 +13,7 @@ export class Transaction {
     })
     note?: string
 
-    @Column('double', {
+    @Column('decimal', {
         nullable: false,
     })
     amount: number
